requirements.txt
alabaster==0.7.12
amqp==2.6.0
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coverage==5.1
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
docutils==0.16
Faker==0.9.1
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
Jinja2==2.11.2
kombu==4.6.10
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.4.0
mypy==0.780
mypy-extensions==0.4.3
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
sentry-sdk==0.14.4
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.2
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.4
whitenoise==5.1.0
wincertstore==0.2
